requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
Django==4.1.6
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
linkedin-api==2.0.0a5
lxml==4.9.2
multidict==6.0.4
openai==0.26.4
pdfkit==1.0.0
phonenumbers==8.13.5
Pillow==9.4.0
psycopg2-binary==2.9.5
pytz==2022.7.1
reportlab==3.6.12
requests==2.28.2
soupsieve==2.3.2.post1
sqlparse==0.4.3
tqdm==4.64.1
tzdata==2022.7
urllib3==1.26.14
yarl==1.8.2
